+++ CHANGELOG.rst
@@ -5,7 +5,8 @@ Changelog
 Version 1.3
 ===========
 
-Changed structure in pyscaffold to import importing of module and version control
+* Changed structure in pyscaffold to import importing of module and version control
+* Included utils section with the *StatLineTable* class to get more structure in the table
 
 Version 1.2
 ===========